Output "lanes" option only when explicitly specified as option.
The "lanes" or "l" (short form) is added to the salt by xgen_salt_argon2() only when

a) Requested explicitly by option
b) Different from ARGON2_MEMORY_LANES default value.

This is because other implementations don't support this, e.g. python's passlib or the argon2 CLI utility.

I don't see it in RFC9106 either, but keep our support to make it possible to specify it as a separate option.