Introduce new GUC pgxcrypto.scrypt_magic.
This allows to configure the magic identifier on password hashes to be either
"$7$" or "$scrypt$", which are used differently on "crypt" and the other supported
backends. A follow up commit will adjust the generated salt string and options, so that
they match this as well.

Includes many code fixes and repairs builds via PGXS.