Check crypt() for invalid hash.
According to crypt(3) documentation/manpage, some implementations might return an invalid hash.
This is recognized by a starting '*' character.
The docs also states that the string is shorter than 13 characters, but it looks like testing for '*' should be enough:
We don't expect the final hash to start with anything different than our supported magic bytes.