Fix some bugs in argon2 and scrypt implementation.
Also make sure we support salt strings without trailing '$'. This revealed a bug in our simple salt parser, as described below:

Our simple salt parser implementation wasn't correctly initialized by scrypt and argon2.
We forgot to properly set the initial salt string length to zero. This caused callers when checking the salt length for proper length check to fail in cases where it was just set randomly during allocation.

Adjust the regression test for scrypt accordingly. Surprisingly this passed just by sheer luck before, since the code checked for the number of parsed sections.
But this isn't real anymore, since we now also support salt strings without a trailing '$', which won't set the number of parsed sections.
This relies on the number of parsed '$', which counts the number of section the simple salt parser has seen. But code should now rely on the correct length of each parsed section, so make sure the behavior is correct.

I was thinking very long if we really need to support salts without the trailing '$', but python's passlib and all others do the same.
So make us compatible here.